Split driver into general and simple parts
The general part is the general driver and no longer even depends on
codecs or channels.

The simple part is the specialisation to codecs and channels and
corresponds to the original interface.

Docs still need to be tidied up after this split.
