+++ typed-protocols/typed-protocols.cabal
@@ -24,6 +24,8 @@ library
                    , Network.TypedProtocol.Channel
                    , Network.TypedProtocol.Codec
                    , Network.TypedProtocol.Driver
+                   , Network.TypedProtocol.Driver.General
+                   , Network.TypedProtocol.Driver.Simple
                    , Network.TypedProtocol.Driver.ByteLimit
                    , Network.TypedProtocol.Proofs
 
@@ -65,6 +67,8 @@ test-suite test-protocols
                    , Network.TypedProtocol.Codec
                    , Network.TypedProtocol.Core
                    , Network.TypedProtocol.Driver
+                   , Network.TypedProtocol.Driver.General
+                   , Network.TypedProtocol.Driver.Simple
                    , Network.TypedProtocol.Driver.ByteLimit
                    , Network.TypedProtocol.PingPong.Client
                    , Network.TypedProtocol.PingPong.Codec
